Improves user model
