+++ src/Blorc.OpenIdConnect/Models/Profile.cs
@@ -1,122 +1,56 @@
 ﻿namespace Blorc.OpenIdConnect
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Text.Json;
+    using System.Text.Json.Serialization;
 
-    [ObsoleteEx(Message = "This will be removed in favor of custom user model", RemoveInVersion = "2.0.0")]
-    public class Profile : IProfile
+    public class Profile
     {
-        private readonly JsonElement _jsonElement;
-
-        private List<string> _roles;
-
-        public Profile(JsonElement jsonElement)
-        {
-            _jsonElement = jsonElement;
-        }
-
+        [JsonPropertyName("email")]
         public string Email
         {
-            get
-            {
-                if (_jsonElement.TryGetProperty("email", out var value))
-                {
-                    return value.GetString();
-                }
-
-                return string.Empty;
-            }
+            get;
+            set;
         }
 
+        [JsonPropertyName("email_verified")]
         public bool EmailVerified
         {
-            get
-            {
-                if (_jsonElement.TryGetProperty("email_verified", out var value))
-                {
-                    return value.GetBoolean();
-                }
-
-                return false;
-            }
+            get;
+            set;
         }
 
+        [JsonPropertyName("family_name")]
         public string FamilyName
         {
-            get
-            {
-                if (_jsonElement.TryGetProperty("family_name", out var value))
-                {
-                    return value.GetString();
-                }
-
-                return string.Empty;
-            }
+            get;
+            set;
         }
 
+        [JsonPropertyName("given_name")]
         public string GivenName
         {
-            get
-            {
-                if (_jsonElement.TryGetProperty("given_name", out var value))
-                {
-                    return value.GetString();
-                }
-
-                return string.Empty;
-            }
+            get;
+            set;
         }
 
-        public JsonElement JsonElement => _jsonElement;
-
+        [JsonPropertyName("name")]
         public string Name
         {
-            get
-            {
-                if (_jsonElement.TryGetProperty("name", out var value))
-                {
-                    return value.GetString();
-                }
-
-                return string.Empty;
-            }
+            get;
+            set;
         }
 
+        [JsonPropertyName("preferred_username")]
         public string PreferredUsername
         {
-            get
-            {
-                if (_jsonElement.TryGetProperty("preferred_username", out var value))
-                {
-                    return value.GetString();
-                }
-
-                return string.Empty;
-            }
+            get;
+            set;
         }
 
+        [JsonPropertyName("roles")]
         public string[] Roles
         {
-            get
-            {
-                if (_roles is null)
-                {
-                    if (_jsonElement.TryGetProperty("roles", out var value))
-                    {
-                        _roles = new List<string>();
-                        using (var elements = value.EnumerateArray())
-                        {
-                            foreach (var element in elements)
-                            {
-                                _roles.Add(element.GetString());
-                            }
-                        }
-                    }
-                }
-
-                return _roles is null ? Array.Empty<string>() : _roles.ToArray();
-            }
+            get;
+            set;
         }
     }
 }
